+++ alerter/test/alerter/alerters/test_system.py
@@ -1403,7 +1403,8 @@ class TestSystemAlerter(unittest.TestCase):
         data_for_alerting = []
         self.test_system_alerter._create_state_for_system(self.system_id)
         self.test_system_alerter._system_initial_alert_sent[
-            self.system_id][GroupedSystemAlertsMetricCode.SystemIsDown.value] = True
+            self.system_id][
+            GroupedSystemAlertsMetricCode.SystemIsDown.value] = True
         data = self.data_received_initially_no_alert['result']['data']
         data['went_down_at']['previous'] = self.last_monitored
         meta_data = self.data_received_initially_no_alert['result']['meta_data']
@@ -1426,7 +1427,8 @@ class TestSystemAlerter(unittest.TestCase):
         self.test_system_alerter._create_state_for_system(self.system_id)
         # Set that the initial downtime alert was sent already
         self.test_system_alerter._system_initial_alert_sent[
-            self.system_id][GroupedSystemAlertsMetricCode.SystemIsDown.value] = True
+            self.system_id][
+            GroupedSystemAlertsMetricCode.SystemIsDown.value] = True
         data = self.data_received_initially_no_alert['result']['data']
         data['went_down_at']['previous'] = self.last_monitored
         meta_data = self.data_received_initially_no_alert['result']['meta_data']
@@ -2594,60 +2596,6 @@ class TestSystemAlerter(unittest.TestCase):
                                                    properties, body)
             mock_create_state_for_system.assert_not_called()
             mock_process_errors.assert_not_called()
-        except Exception as e:
-            self.fail("Test failed: {}".format(e))
-
-    @mock.patch(
-        "src.alerter.alerters.system.SystemAlerter._place_latest_data_on_queue")
-    @mock.patch("src.alerter.alerters.system.SystemAlerter._process_results")
-    @mock.patch(
-        "src.alerter.alerters.system.SystemAlerter._create_state_for_system")
-    @mock.patch.object(RabbitMQApi, "basic_ack")
-    def test_process_result_data_send_data_called(
-            self, mock_ack, mock_create_state_for_system,
-            mock_process_results, mock_place_latest_data_on_queue) -> None:
-
-        mock_ack.return_value = None
-        mock_create_state_for_system.return_value = None
-        mock_process_results.return_value = None
-
-        try:
-            self.test_system_alerter.rabbitmq.connect()
-            blocking_channel = self.test_system_alerter.rabbitmq.channel
-            method = pika.spec.Basic.Deliver(
-                routing_key=self.input_routing_key)
-            body = json.dumps(self.data_received_initially_no_alert)
-            properties = pika.spec.BasicProperties()
-            self.test_system_alerter._process_data(blocking_channel, method,
-                                                   properties, body)
-            mock_place_latest_data_on_queue.assert_called_once()
-        except Exception as e:
-            self.fail("Test failed: {}".format(e))
-
-    @mock.patch(
-        "src.alerter.alerters.system.SystemAlerter._place_latest_data_on_queue")
-    @mock.patch("src.alerter.alerters.system.SystemAlerter._process_errors")
-    @mock.patch(
-        "src.alerter.alerters.system.SystemAlerter._create_state_for_system")
-    @mock.patch.object(RabbitMQApi, "basic_ack")
-    def test_process_error_data_send_data_called(
-            self, mock_ack, mock_create_state_for_system,
-            mock_process_errors, mock_place_latest_data_on_queue) -> None:
-
-        mock_ack.return_value = None
-        mock_create_state_for_system.return_value = None
-        mock_process_errors.return_value = None
-
-        try:
-            self.test_system_alerter.rabbitmq.connect()
-            blocking_channel = self.test_system_alerter.rabbitmq.channel
-            method = pika.spec.Basic.Deliver(
-                routing_key=self.input_routing_key)
-            body = json.dumps(self.data_received_error_data)
-            properties = pika.spec.BasicProperties()
-            self.test_system_alerter._process_data(blocking_channel, method,
-                                                   properties, body)
-            mock_place_latest_data_on_queue.assert_called_once()
         except Exception as e:
             self.fail("Test failed: {}".format(e))