+++ docs/CHANGE_LOG.md
@@ -19,6 +19,7 @@ Change the contents of this file to this:
 - The data store components are now compatible with the base Chainlink integration features.
 - Added Chainlink Node Alerter Manager and tests.
 - Web-Installer bug fixes and removing outdated alerts.
+- Added Chainlink Node Alerter logic and tests.
 
 ## 0.1.2